Rename helper function folder, files and references
