+++ frontend/helpers/supportHelpers.js
